Moved connections under project.
1. Additionally created the new migrations by removing all
2. separate sub pages within project_detail to manage members, desc and db connections
