--- connection/fixtures/db_type.yaml
@@ -1,12 +1,0 @@
-- model: connection.DbType
-  pk: 1
-  fields:
-    dbname: 'PostgreSQL'
-- model: connection.dbtype
-  pk: 2
-  fields:
-    dbname: 'MySQL'
-- model: connection.dbtype
-  pk: 3
-  fields:
-    dbname: 'MSSQL'
