+++ data_compare/templates/account/signup.html
@@ -38,10 +38,10 @@
           <div class="field">
             <div class="helptext has-text-left content"  id="id_password1_helptext">
               <ul class="is-size-7">
-                <li>Your password can’t be too similar to your other personal information.</li>
+                <li>Your password can't be too similar to your other personal information.</li>
                 <li>Your password must contain at least 8 characters.</li>
-                <li>Your password can’t be a commonly used password.</li>
-                <li>Your password can’t be entirely numeric.</li>
+                <li>Your password can't be a commonly used password.</li>
+                <li>Your password can't be entirely numeric.</li>
               </ul>
             </div>
             <div class="control">
