+++ data_compare/templates/app_base.html
@@ -67,7 +67,6 @@
               <div class="navbar-item has-dropdown is-hoverable">
                 <a class="navbar-link">Go To</a>
                 <div class="navbar-dropdown is-right">
-                  <a class="navbar-item"href="{% url 'settings:connections' %}">{% translate "Connection" %}</a>
                   <a class="navbar-item" href="{% url 'projects' %}">Projects</a>
                   <a class="navbar-item">Components</a>
                   <hr class="navbar-divider" />
