+++ data_compare/users/tests/test_urls.py
@@ -4,8 +4,8 @@ from data_compare.users.models import User
 
 
 def test_detail(user: User):
-    assert reverse("users:detail", kwargs={"pk": user.pk}) == f"/users/{user.pk}/"
-    assert resolve(f"/users/{user.pk}/").view_name == "users:detail"
+    assert reverse("users:user_detail", kwargs={"pk": user.pk}) == f"/users/{user.pk}/"
+    assert resolve(f"/users/{user.pk}/").view_name == "users:user_detail"
 
 
 def test_update():
