+++ project/admin.py
