+++ project/templates/connection_form.html
@@ -9,7 +9,7 @@
         </br />
         <div class="columns">
           <div class="column">
-            <div class="field">
+            <div class="field is-small">
               <label class="label" for="id_dbname">Name</label>
               <div class="control">
                 <input value="{{ connection.name }}"
@@ -106,6 +106,7 @@
                        class="input"
                        placeholder="Port No." />
               </div>
+              {% if form.portno.errors %}<p class="help is-danger">{{ form.portno.errors|striptags }}</p>{% endif %}
             </div>
           </div>
           <div class="column">
@@ -127,7 +128,7 @@
           <button type="submit"
                   name="test"
                   class="button is-info"
-                  hx-post="{% url 'settings:create_connection' %}"
+                  hx-post="{% url 'create_connection' 1 %}"
                   hx-swap="outerhtml"
                   hx-trigger="click"
                   hx-target="#connection_result">Test Connection</button>
