+++ project/templates/connection_index.html
@@ -1,8 +1,3 @@
-{% extends "app_base.html" %}
-
-{% block title %}
-  Database Connections
-{% endblock title %}
 {% block content %}
   {% load static %}
 
@@ -10,20 +5,17 @@
     <nav class="breadcrumb" aria-label="breadcrumbs">
       <ul>
         <li>
-          <a href="#">Settings</a>
-        </li>
-        <li>
           <a href="#">Database Connections</a>
         </li>
         <li>
-          <a hx-get="{% url 'settings:create_connection' %}" hx-target="#dialog">
+          <a hx-get="{% url 'create_connection' project.id %}" hx-target="#dialog">
             <i class="fa-solid fa-plus fa-lg has-text-success"></i>
           </a>
         </li>
       </ul>
     </nav>
     <div hx-trigger="load, listChanged from:body"
-         hx-get="{% url 'settings:list_connection' 1 %}"
+         hx-get="{% url 'list_connection' project.id 1 %}"
          hx-target="this"
          id="connection_items">
       <tr class="text-center">
@@ -34,7 +26,4 @@
     </div>
   </div>
   <!-- Placeholder for the modal -->
-  <div id="modal" class="modal">
-    <div id="dialog" class="modal-dialog" hx-target="this"></div>
-  </div>
 {% endblock content %}
