+++ project/templates/partials/project_delete.html
@@ -1,0 +1,15 @@
+<div class="pt-5 pb-5 is-size-4 has-text-danger">Danger Zone</div>
+<span class="pt-5 pb-2">
+  Tread carefully here. You delete your project, you lose all your data,
+  your members can't access projects and pages, and we can't retrieve any of it for you.
+  Proceed only if you are sure you want your project deleted.
+</span>
+<p class="pt-5">
+  <a class="button is-danger"
+     hx-post="{% url 'delete_project' project.id %}"
+     hx-confirm="Are you sure you wish to delete your {{ object.name }}?"
+     hx-headers='{"X-CSRFToken":"{{ csrf_token }}"}'>
+    <i class="fa fa-trash fa-fw" aria-hidden="true"></i>
+    Delete {{ object.name }} Project
+  </a>
+</p>
