+++ project/templates/partials/project_members.html
@@ -1,0 +1,69 @@
+{% block content %}
+  <div class="container">
+    <div class="columns">
+      <div class="column is-6">
+        <h3 class="is-size-5">{{ project.name }} Members</h3>
+      </div>
+      <div class="column">
+        <form hx-post="{% url 'project_members' project.id %}?action=add"
+              method="post"
+              hx-target="#member-table"
+              hx-swap="beforeend">
+          {% csrf_token %}
+          <div class="field is-grouped">
+            <p class="control is-expanded">
+              <input class="input is-small"
+                     id="id_email"
+                     name="email"
+                     type="email"
+                     placeholder="Add Email address" />
+            </p>
+            <p class="control">
+              <button class="button is-info is-small">Add Member</button>
+            </p>
+          </div>
+        </form>
+      </div>
+    </div>
+    <div class="container pt-5">
+      <table class="table is-narrow is-hoverable is-fullwidth">
+        <thead>
+          <tr>
+            <th>Full Name</th>
+            <th>Email</th>
+            <th>Joined On</th>
+            <th>Role</th>
+            <th>Action</th>
+          </tr>
+        </thead>
+        <tbody id="member-table">
+          {% for member in members %}
+            <tr class="small" id="member-{{ member.id }}">
+              <td class="is-vcentered">{{ member.name }}</td>
+              <td class="is-vcentered">{{ member.email }}</td>
+              <td class="is-vcentered">{{ member.date_joined|date:"d-M-y" }}</td>
+              <td class="is-vcentered">
+                <div class="select">
+                  <select class="is-small">
+                    <option>Member</option>
+                    <option>Admin</option>
+                  </select>
+                </div>
+              </td>
+              <td class="has-text-danger is-vcentered">
+                <button class="has-text-danger"
+                        hx-post="/project/{{ project.id }}/project_members?action=remove&user_id={{ member.id }}"
+                        hx-headers='{"X-CSRFToken":"{{ csrf_token }}"}'
+                        hx-target="#member-{{ member.id }}"
+                        hx-confirm="Are you sure about removing your member: '{{ member.email }}?'"
+                        hx-swap="outerHTML:remove">
+                  <i class="fa-solid fa-trash"></i>
+                </button>
+              </td>
+            </tr>
+          {% endfor %}
+        </tbody>
+      </table>
+    </div>
+  </div>
+{% endblock content %}
