+++ project/templates/project_confirm_delete.html
