+++ project/templates/project_detail.html
@@ -1,0 +1,116 @@
+{% extends "app_base.html" %}
+
+{% block title %}
+  {{ project.name }}
+{% endblock title %}
+{% block content %}
+  {% load static %}
+
+  <div class="container">
+    <nav class="breadcrumb" aria-label="breadcrumbs">
+      <ul>
+        <li>
+          <a class="pl-3 has-text-weight-bold is-size-5 has-text-dark"
+             href="{% url 'project_details' project.id %}">{{ project.name }} Configurations</a>
+        </li>
+      </ul>
+    </nav>
+    <div class="columns">
+      <div class="column is-3">
+        <ul class="menu menu-list">
+          <li>
+            <a class="has-text-weight-bold is-active"
+               hx-get="{% url 'configure_edit_project' object.id %}"
+               hx-target="#config_content"
+               hx-swap="innerHTML"
+               hx-on:click="setActive(this)">General</a>
+          </li>
+          <li>
+            <a class="has-text-weight-bold"
+               hx-get="{% url 'project_members' object.id %}"
+               hx-target="#config_content"
+               hx-swap="innerHTML"
+               hx-on:click="setActive(this)">Members</a>
+          </li>
+          <li>
+            <a class="has-text-weight-bold"
+               hx-get="{% url 'connections' object.id %}"
+               hx-target="#config_content"
+               hx-swap="innerHTML"
+               hx-on:click="setActive(this)">Connection Settings</a>
+          </li>
+        </ul>
+      </div>
+      <div class="column is-9" id="side-view">
+        <div id="config_content">
+          <h3 class="is-size-4">{{ object.name }} project</h3>
+          <form hx-post="{% url 'configure_edit_project' object.id %}"
+                hx-headers='{"X-CSRFToken":"{{ csrf_token }}"}'>
+            <div>
+            </br />
+            <div class="columns">
+              <div class="column is-8">
+                <div class="field">
+                  <label class="label" for="id_name">Project name</label>
+                  <div class="control">
+                    <input value="{{ object.name }}"
+                           type="text"
+                           name="name"
+                           id="id_name"
+                           class="input is-small"
+                           required
+                           placeholder="Project Name" />
+                  </div>
+                </div>
+              </div>
+              <div class="column">
+                <div class="field">
+                  <label class="label" for="id_project_code">Project Code</label>
+                  <div class="control">
+                    <input value="{{ object.project_code }}"
+                           type="text"
+                           name="project_code"
+                           id="id_project_code"
+                           class="input is-small"
+                           required
+                           placeholder="Project Code" />
+                  </div>
+                </div>
+              </div>
+            </div>
+            <div class="columns">
+              <div class="column">
+                <div class="field">
+                  <label class="label" for="id_description">Project Description</label>
+                  <textarea id="id_description"
+                            name="description"
+                            class="textarea is-small"
+                            placeholder="e.g. Describe briefly about the project...">{{ object.description }}</textarea>
+                </div>
+                {% if form.description.errors %}<p class="help is-danger">{{ form.description.errors|striptags }}</p>{% endif %}
+              </div>
+            </div>
+            <div class="buttons">
+              <button class="button is-primary" name="save">Save</button>
+            </div>
+          </div>
+        </form>
+        {% include 'partials/project_delete.html' %}
+      </div>
+    </div>
+  </div>
+</div>
+<!-- Placeholder for the modal -->
+<div id="modal" class="modal modal-lg fade">
+  <div id="dialog" class="modal-dialog" hx-target="this"></div>
+</div>
+{% endblock content %}
+{% block inline_javascript %}
+  <script>
+    function setActive(element) {
+      document.querySelectorAll('a').forEach(link =>
+        link.classList.remove('is-active'));
+      element.classList.add('is-active');
+    }
+  </script>
+{% endblock inline_javascript %}
