+++ project/templates/project_index.html
