--- testplan/templates/project_detail.html
@@ -1,22 +1,0 @@
-{% extends "app_base.html" %}
-
-{% block title %}
-  {{ project.name }}
-{% endblock title %}
-{% block content %}
-  {% load static %}
-
-  <div class="container">
-    <nav class="breadcrumb" aria-label="breadcrumbs">
-      <ul>
-        <li>
-          <a href="{% url 'projects' %}">{{ project.name }}</a>
-        </li>
-      </ul>
-    </nav>
-  </div>
-  <!-- Placeholder for the modal -->
-  <div id="modal" class="modal modal-lg fade">
-    <div id="dialog" class="modal-dialog" hx-target="this"></div>
-  </div>
-{% endblock content %}
