+++ testplan/urls.py
@@ -3,13 +3,6 @@ from django.urls import path
 from testplan import views
 
 urlpatterns = [
-    # Project URLs
-    path("projects", views.project_index, name="projects"),
-    path("project/list/<int:page>", views.ProjectListView.as_view(), name="list_project"),
-    path("project/create", views.ProjectCreateView.as_view(), name="create_project"),
-    path("project/<int:pk>/edit", views.ProjectUpdateView.as_view(), name="edit_project"),
-    path("project/<int:pk>/delete", views.ProjectDeleteView.as_view(), name="delete_project"),
-    path("project/<int:pk>", views.ProjectDetailView.as_view(), name="project_details"),
     # Test Case URLs
     path("project/<int:project_id>/testcase/create", views.TestCaseCreateView.as_view(), name="create_testcase"),
     path("project/<int:project_id>/testcases", views.TestCaseListView.as_view(), name="list_testcase"),
