+++ testplan/views.py
@@ -5,7 +5,6 @@ import pandas as pd
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
 from django.core.exceptions import PermissionDenied
-from django.core.paginator import Paginator
 from django.db.models.aggregates import Count
 from django.db.models.deletion import ProtectedError
 from django.http import HttpResponse
@@ -13,199 +12,13 @@ from django.http.response import HttpResponseRedirect
 from django.shortcuts import get_object_or_404, redirect, render
 from django.views.generic import CreateView, DeleteView, DetailView, ListView, UpdateView
 
-from connection.models import DbConnection
+from project.models import DbConnection
 from testplan import tasks
 from testplan.models import Project, TestCase, TestRun, TestRunTestCase, TestRunTestCaseHistory
 
 _404_Page = "404.html"
 
 logger = logging.getLogger(__name__)
-
-
-@login_required
-def project_index(request):
-    logger.info("Rendering Project Index")
-    return render(request, "project_index.html")
-
-
-class ProjectCreateView(LoginRequiredMixin, CreateView):
-    model = Project
-    fields = ["name", "project_code", "description"]
-    template_name = "project_form.html"
-
-    def form_valid(self, form):
-        project = form.save(commit=False)
-        project.owner = self.request.user
-        project.created_by = self.request.user
-        project.save()
-        form.save_m2m()
-        project.members.add(self.request.user.id)
-        return HttpResponse(
-            status=204,
-            headers={
-                "HX-Trigger": json.dumps(
-                    {
-                        "listChanged": None,
-                        "showMessage": f"{project.name} added.",
-                        "eventType": "created",
-                    }
-                )
-            },
-        )
-
-    def form_invalid(self, form):
-        return self.render_to_response(self.get_context_data(form=form))
-
-
-class ProjectUpdateView(LoginRequiredMixin, UserPassesTestMixin, UpdateView):
-    model = Project
-    fields = ["name", "project_code", "description"]
-    template_name = "project_form.html"
-
-    def form_valid(self, form):
-        project = form.save(commit=False)
-        project.modified_by = self.request.user
-        project.save()
-        form.save_m2m()
-        return HttpResponse(
-            status=204,
-            headers={
-                "HX-Trigger": json.dumps(
-                    {
-                        "listChanged": None,
-                        "showMessage": f"{project.name} updated.",
-                        "eventType": "updated",
-                    }
-                )
-            },
-        )
-
-    def form_invalid(self, form):
-        return self.render_to_response(self.get_context_data(form=form))
-
-    def get_success_message(self, cleaned_data):
-        name = cleaned_data["name"]
-        return self.success_message % dict(
-            {
-                "cleaned_data": cleaned_data,
-                "name": name,
-                "object_id": self.kwargs["project_id"],
-            }
-        )
-
-    def get_success_url(self):
-        return redirect("projects")
-
-    def test_func(self):
-        if self.request.user == self.get_object().owner:
-            return True
-        return False
-
-    def handle_no_permission(self):
-        project = self.get_object()
-        return HttpResponse(
-            status=204,
-            headers={
-                "HX-Trigger": json.dumps(
-                    {
-                        "showMessage": f"You do not have permission to edit '{project.name}'",
-                        "eventType": "permissiondenied",
-                    }
-                )
-            },
-        )
-
-
-class ProjectListView(LoginRequiredMixin, ListView):
-    model = Project
-    template_name = "project_list.html"
-
-    def get_context_data(self, **kwargs):
-        logger.info("Came in ProjectListView.get_context_data")
-        context = super().get_context_data(**kwargs)
-        page_num = self.kwargs.get("page")
-        member_projects = (
-            Project.objects.filter(members__id__exact=self.request.user.id).order_by("-owner").select_related("owner")
-        )
-        paginator = Paginator(member_projects, per_page=5)
-        context["projects"] = paginator.get_page(page_num)
-        return context
-
-
-class ProjectDeleteView(LoginRequiredMixin, UserPassesTestMixin, DeleteView):
-    model = Project
-    template_name = "project_confirm_delete.html"
-
-    def get_object(self, queryset=None):
-        project = Project.objects.get(id=self.kwargs.get("pk"))
-        return project
-
-    def post(self, *args, **kwargs):
-        project = self.get_object()
-        try:
-            self.get_object().delete()
-            return HttpResponse(
-                status=204,
-                headers={
-                    "HX-Trigger": json.dumps(
-                        {
-                            "listChanged": None,
-                            "showMessage": f"{project.name} deleted.",
-                            "eventType": "deleted",
-                        }
-                    )
-                },
-            )
-        except ProtectedError as e:
-            return HttpResponse(
-                status=204,
-                headers={
-                    "HX-Trigger": json.dumps(
-                        {
-                            "listChanged": None,
-                            "showMessage": "'{}' cannot be deleted as it has {} dependent records".format(
-                                project.name, len(e.args[1])
-                            ),
-                            "eventType": "deleted",
-                        }
-                    )
-                },
-            )
-
-    def test_func(self):
-        project = self.get_object()
-        if project.owner == self.request.user:
-            return True
-        return False
-
-    def handle_no_permission(self):
-        project = self.get_object()
-        if self.raise_exception:
-            raise PermissionDenied(self.get_permission_denied_message())
-        return HttpResponse(
-            status=204,
-            headers={
-                "HX-Trigger": json.dumps(
-                    {
-                        "showMessage": f"You do not have permission to delete '{project.name}'",
-                        "eventType": "permissiondenied",
-                    }
-                )
-            },
-        )
-
-
-class ProjectDetailView(LoginRequiredMixin, UserPassesTestMixin, DetailView):
-    model = Project
-    template_name = "project_detail.html"
-
-    def test_func(self):
-        if self.request.user in self.get_object().members.all():
-            return True
-        return False
-
-    def handle_no_permission(self) -> HttpResponseRedirect:
-        return render(self.request, _404_Page)
 
 
 class TestCaseCreateView(LoginRequiredMixin, UserPassesTestMixin, CreateView):
